book_rental_manager/requirements.txt
flask
flask_pymongo
flask_restx
faker
factory_boy
flask_cors
mongoengine
mongomock
mongoengine-goodjson